Removing unhelpful documentation
@haseebjkhan I don't see how this info is helpful. I can see the sizes below.

Mentioning other stuff like the icons are designed on consistent canvas size that is slightly bigger than the icon might be helpful.

Or product useage:
Outlined icons primaryily used for…

Anyways, this sort of documentated can be added later. We need to focus on the content and assets, not notes.